Potential way to set up for ROS2 Foxy
Deferring this until later. Going to start on Noetic and figure out how (or if) docker will play a role in this. Want to start with ROS1 because I know how to use that rather than jumping into the deep end with docker and ROS2 (which is fairly undocumented). Upgrade to ROS2 can be easier once ROS1 code is made